Refactor configs to be more generic
