+++ app/schemas/credential_registry/credential/config.json
