+++ app/schemas/credential_registry/organization/config.json
