+++ app/schemas/learning_registry/config.json
