+++ spec/api/v1/envelopes_spec.rb
@@ -182,7 +182,7 @@ describe API::V1::Envelopes do
 
       it 'returns a 201 Created http status code' do
         publish.call
-
+        binding.pry
         expect_status(:created)
       end
